Moved function to its own class for better organization.
